LayoutTests/fast/page-color-sampling/color-sampling-fixed-ancestor-with-relative-container.html

<!DOCTYPE html> <!-- webkit-test-runner [ ContentInsetBackgroundFillEnabled=true pageTopColorSamplingEnabled=true useFlexibleViewport=true obscuredInset.top=100 ] -->
<html>
<head>
    <meta charset="utf-8">
    <meta name="viewport" content="width=device-width, initial-scale=1">
    <style>
        body, html {
            width: 100%;
            height: 100%;
            margin: 0;
            font-family: system-ui;
        }

        header {
            position: fixed;
            top: -40px;
            width: 100%;
            height: 40px;
        }

        .header-content {
            background: rgb(212, 214, 185);
            width: 100%;
            height: 64px;
            position: absolute;
            top: 40px;
        }
    </style>
    <script src="../../resources/js-test.js"></script>
    <script src="../../resources/ui-helper.js"></script>
    <script>
    jsTestIsAsync = true;

    addEventListener("load", async () => {
        await UIHelper.setObscuredInsets(100, 0, 0, 0);
        await UIHelper.ensurePresentationUpdate();
        edgeColors = await UIHelper.fixedContainerEdgeColors();
        shouldBeEqualToString("edgeColors.top", "rgb(212, 214, 185)");
        finishJSTest();
    });
    </script>
</head>
<body>
<header>
    <div class="header-content"></div>
</header>
<div class="tall"></div>
</body>
</html>
