LayoutTests/fast/page-color-sampling/color-sampling-fixed-ancestor-with-relative-container-expected.txt

PASS edgeColors.top is "rgb(212, 214, 185)"
PASS successfullyParsed is true

TEST COMPLETE
